requirements.txt
urllib3
requests
requests[socks]
pysocks
beautifulsoup4
lxml
python-dotenv